# kvs: improve performance of transaction prep/check