Per discussion in #3633.

I went back and forth on implementation eventually settling on a `zlist` "work_queue".  I dislike the fact I have a "on_work_queue" flag placed in the `struct kvsroot` structure.  It sort of breaks abstraction of that struct.  I could ...

1) just search the list instead of maintaining a flag, but that's a lot of iteration and we don't want to do that.

2) use a `zhash_t` instead, but a `zhash` foreach iterates through all hash buckets, and we're trying to limit iteration

3) use a `zhash_t` along w/ a `zlist_t`, using the `zhash_t` just for "is it on the list" management.  Seems like a lot of effort for just the flag.

Using @grondo's namespace benchmark.

Before

```
  0.19s: Benchmark on 256 namespaces complete: 1382.4 namespace/s
  0.35s: Benchmark on 512 namespaces complete: 1481.6 namespace/s
  0.87s: Benchmark on 1024 namespaces complete: 1179.5 namespace/s
  2.64s: Benchmark on 2048 namespaces complete: 776.0 namespace/s
 12.99s: Benchmark on 4096 namespaces complete: 315.3 namespace/s
 48.11s: Benchmark on 8192 namespaces complete: 170.3 namespace/s
174.14s: Benchmark on 16384 namespaces complete: 94.1 namespace/s
540.87s: Benchmark on 32768 namespaces complete: 60.6 namespace/s
```

After

```
  0.18s: Benchmark on 256 namespaces complete: 1458.1 namespace/s
  0.35s: Benchmark on 512 namespaces complete: 1464.3 namespace/s
  0.72s: Benchmark on 1024 namespaces complete: 1416.8 namespace/s
  1.41s: Benchmark on 2048 namespaces complete: 1450.0 namespace/s
  2.82s: Benchmark on 4096 namespaces complete: 1451.9 namespace/s
  5.61s: Benchmark on 8192 namespaces complete: 1459.7 namespace/s
 11.30s: Benchmark on 16384 namespaces complete: 1449.4 namespace/s
 22.47s: Benchmark on 32768 namespaces complete: 1458.0 namespace/s
```