@chu11 2021-05-14
> Your comment about the list flag, which is presumably to avoid the O(N) zlist_remove() when unnecessary, reminded me that you could switch that list over to the CCAN list I just pulled in. It has the following advantages:

Ahh, I didn't think about the new ccan list you brought in.  What I really need is a O(1) `list_find()` .. which of course doesn't exist, which is why I considered using a `zhash` instead.  I need it to make sure I don't add the same KVS root onto the `work_queue` twice.

But the O(1) remove is appealing, as we'll avoid some extra unnecessary looping.  Let me try that out and see if that helps more.