@garlick 2021-05-14
> I need it to make sure I don't add the same KVS root onto the work_queue twice.

It would be O(1) to test if the `struct list_node` is on any list.  There isn't a macro for it that I can see but you would just check for the `list_node_init()` condition which is `node->next == node && node->prev == node`.