@chu11 2021-05-14
> It would be O(1) to test if the struct list_node is on any list. There isn't a macro for it that I can see but you would just check for the list_node_init() condition which is node->next == node && node->prev == node.

I had just realized it :-)  But yes, doesn't appear there's a macro.  And I think I'd have to use `list_del_init_()` after its removed.