@chu11 2021-05-14
re-pushed using ccan list implementation.  Used the nsbench benchmark at larger sizes under the assumption it'd be hard to see the gain at smaller namespace counts.

`>for i in 256 512 1024 2048 4096 8192 16384 32768 65536 131072 262144; do src/cmd/flux start ./nsbench.py -c $i; done`

Before

```
  0.19s: Benchmark on 256 namespaces complete: 1344.8 namespace/s
  0.35s: Benchmark on 512 namespaces complete: 1467.5 namespace/s
  0.72s: Benchmark on 1024 namespaces complete: 1422.1 namespace/s
  1.40s: Benchmark on 2048 namespaces complete: 1465.1 namespace/s
  2.93s: Benchmark on 4096 namespaces complete: 1399.1 namespace/s
  5.51s: Benchmark on 8192 namespaces complete: 1486.6 namespace/s
 11.56s: Benchmark on 16384 namespaces complete: 1417.4 namespace/s
 23.40s: Benchmark on 32768 namespaces complete: 1400.5 namespace/s
 47.17s: Benchmark on 65536 namespaces complete: 1389.5 namespace/s
124.65s: Benchmark on 131072 namespaces complete: 1051.5 namespace/s
433.20s: Benchmark on 262144 namespaces complete: 605.1 namespace/s
```

After

```
  0.18s: Benchmark on 256 namespaces complete: 1444.1 namespace/s
  0.36s: Benchmark on 512 namespaces complete: 1434.6 namespace/s
  0.68s: Benchmark on 1024 namespaces complete: 1502.4 namespace/s
  1.31s: Benchmark on 2048 namespaces complete: 1564.3 namespace/s
  2.69s: Benchmark on 4096 namespaces complete: 1521.3 namespace/s
  5.54s: Benchmark on 8192 namespaces complete: 1478.7 namespace/s
 11.04s: Benchmark on 16384 namespaces complete: 1484.2 namespace/s
 22.57s: Benchmark on 32768 namespaces complete: 1451.7 namespace/s
 46.86s: Benchmark on 65536 namespaces complete: 1398.6 namespace/s
120.06s: Benchmark on 131072 namespaces complete: 1091.7 namespace/s
386.44s: Benchmark on 262144 namespaces complete: 678.4 namespace/s
```

I only ran once, but at the larger sizes we can see the improvement.