@grondo 2021-05-14
> @chu11: should I try your branch for my reproducer with Fluxion at this point?

@dongahn (in case @chu11 is away from computer), I would definitely try out this PR if you have time. Though we have high confidence that the two referenced benchmarks accurately simulate the original issue, positive verification @chu11's fixes help your specific test will be very valuable!