@grondo 2021-05-14
BTW, I anticipate the next bottleneck will be in the job-exec module while it tries to keep up with the launch of 24K job shells in just a couple minutes.